# docs/openapi-specification.yaml
openapi: 3.0.3
info:
  title: Thai Auto Insurance API
  version: 1.0.0
  description: |
    Comprehensive REST API for Thai Auto Insurance application supporting policy management, claims processing, and customer service operations.
    
    ## Features
    - Multi-language support (Thai/English)
    - JWT-based authentication
    - Thai-specific validations (National ID, phone numbers, addresses)
    - Real-time payment processing
    - Comprehensive claims management
    - Regulatory compliance (OIC requirements)
    
  contact:
    name: API Support Team
    email: api-support@autoinsurance.co.th
    url: https://docs.autoinsurance.co.th
  license:
    name: MIT License
    url: https://opensource.org/licenses/MIT
  termsOfService: https://autoinsurance.co.th/terms
servers:
  - url: https://api.autoinsurance.co.th/api/v1
    description: Production server
  - url: https://api-staging.autoinsurance.co.th/api/v1
    description: Staging server
  - url: https://api-dev.autoinsurance.co.th/api/v1
    description: Development server
security:
  - BearerAuth: []
...
paths:
...
  /policies/quote:
    post:
      tags: [Policy Management]
      summary: Generate insurance quote
      description: Calculate insurance premium quote for vehicle
      operationId: generateQuote
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QuoteRequest'
      responses:
        '200':
          description: Quote generated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuoteResponse'
  /policies:
    get:
      tags: [Policy Management]
      summary: Get customer policies
      description: Retrieve all policies for current customer
      operationId: getCustomerPolicies
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [ACTIVE, EXPIRED, CANCELLED]
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Policies retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PolicyListResponse'

    post:
      tags: [Policy Management]
      summary: Purchase insurance policy
      description: Purchase policy from generated quote
      operationId: purchasePolicy
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PurchasePolicyRequest'
      responses:
        '201':
          description: Policy purchased successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PurchasePolicyResponse'
...
  /policies/{policyId}/renew:
    put:
      tags: [Policy Management]
      summary: Renew policy
      description: Renew existing policy
      operationId: renewPolicy
      parameters:
        - name: policyId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RenewPolicyRequest'
      responses:
        '200':
          description: Policy renewed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RenewPolicyResponse'
  /policies/{policyId}:
    delete:
      tags: [Policy Management]
      summary: Cancel policy
      description: Cancel existing policy
      operationId: cancelPolicy
      parameters:
        - name: policyId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CancelPolicyRequest'
      responses:
        '200':
          description: Policy cancelled successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CancelPolicyResponse'

  # Claims Management APIs
...
  /admin/policies:
    get:
      tags: [Admin Operations]
      summary: Get all policies (Admin)
      description: Retrieve paginated list of all policies
      operationId: getAllPolicies
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 50
        - name: status
          in: query
          schema:
            type: string
            enum: [ACTIVE, EXPIRED, CANCELLED]
      responses:
        '200':
          description: Policies retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminPolicyListResponse'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from login endpoint
  schemas:
...
    PersonalInfo:
      type: object
      required:
        - firstName
        - lastName
        - nationalId
        - dateOfBirth
        - gender
      properties:
        firstName:
          type: string
          example: "สมชาย"
        lastName:
          type: string
          example: "ใจดี"
        firstNameEn:
          type: string
          example: "Somchai"
        lastNameEn:
          type: string
          example: "Jaidee"
        nationalId:
          type: string
          pattern: '^[0-9]{13}$'
          example: "1234567890123"
        dateOfBirth:
          type: string
          format: date
          example: "1990-01-15"
        gender:
          type: string
          enum: [MALE, FEMALE, OTHER]
          example: "MALE"
...
    QuoteRequest:
      type: object
      required:
        - vehicleId
        - quoteRequest
      properties:
        customerId:
          type: string
          format: uuid
        vehicleId:
          type: string
          format: uuid
        quoteRequest:
          $ref: '#/components/schemas/QuoteRequestDetails'
    QuoteRequestDetails:
      type: object
      required:
        - policyType
        - policyPeriod
      properties:
        policyType:
          type: string
          enum: [CMI, CLASS_1, CLASS_2, CLASS_3]
          example: "CLASS_1"
        coverageOptions:
          $ref: '#/components/schemas/CoverageOptions'
        discounts:
          $ref: '#/components/schemas/DiscountOptions'
        policyPeriod:
          $ref: '#/components/schemas/PolicyPeriod'
    CoverageOptions:
      type: object
      properties:
        deductible:
          type: number
          format: float
          example: 5000.00
        medicalExpenses:
          type: number
          format: float
          example: 100000.00
        personalAccident:
          type: number
          format: float
          example: 200000.00
        bailBond:
          type: number
          format: float
          example: 50000.00
        propertyDamage:
          type: number
          format: float
          example: 1000000.00
    DiscountOptions:
      type: object
      properties:
        noClaimDiscount:
          type: boolean
          example: true
        multiVehicleDiscount:
          type: boolean
          example: false
        loyaltyDiscount:
          type: boolean
          example: true
    PolicyPeriod:
      type: object
      required:
        - startDate
        - endDate
        - periodMonths
      properties:
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        periodMonths:
          type: integer
          minimum: 1
          maximum: 12
          example: 12
    QuoteResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          properties:
            quoteId:
              type: string
              format: uuid
            quoteNumber:
              type: string
              example: "QT-2024-123456"
            validUntil:
              type: string
              format: date-time
            premium:
              $ref: '#/components/schemas/PremiumBreakdown'
            coverage:
              $ref: '#/components/schemas/CoverageDetails'
            paymentOptions:
              type: array
              items:
                $ref: '#/components/schemas/PaymentOption'
        timestamp:
          type: string
          format: date-time
    PremiumBreakdown:
      type: object
      properties:
        basePremium:
          type: number
          format: float
          example: 15000.00
        discounts:
          type: object
          properties:
            noClaimDiscount:
              type: number
              format: float
              example: -1500.00
            loyaltyDiscount:
              type: number
              format: float
              example: -500.00
            totalDiscounts:
              type: number
              format: float
              example: -2000.00
        fees:
          type: object
          properties:
            oicFund:
              type: number
              format: float
              example: 50.00
            stamp:
              type: number
              format: float
              example: 1.00
            vat:
              type: number
              format: float
              example: 910.00
            totalFees:
              type: number
              format: float
              example: 961.00
        netPremium:
          type: number
          format: float
          example: 13000.00
        totalPremium:
          type: number
          format: float
          example: 13961.00
    CoverageDetails:
      type: object
      properties:
        policyType:
          type: string
          example: "CLASS_1"
        deductible:
          type: number
          format: float
        coverageDetails:
          type: object
          properties:
            thirdPartyLiability:
              type: object
              properties:
                bodilyInjuryPerPerson:
                  type: number
                  format: float
                bodilyInjuryPerAccident:
                  type: number
                  format: float
                propertyDamage:
                  type: number
                  format: float
            ownDamage:
              type: object
              properties:
                vehicleValue:
                  type: number
                  format: float
                deductible:
                  type: number
                  format: float
            medicalExpenses:
              type: number
              format: float
            personalAccident:
              type: number
              format: float
    PaymentOption:
      type: object
      properties:
        type:
          type: string
          enum: [FULL_PAYMENT, INSTALLMENT_6, INSTALLMENT_12]
        amount:
          type: number
          format: float
        discount:
          type: number
          format: float
        finalAmount:
          type: number
          format: float
        monthlyAmount:
          type: number
          format: float
        totalAmount:
          type: number
          format: float
        interestRate:
          type: number
          format: float
    PolicyListResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          properties:
            policies:
              type: array
              items:
                $ref: '#/components/schemas/PolicySummaryExtended'
            pagination:
              $ref: '#/components/schemas/Pagination'
        timestamp:
          type: string
          format: date-time
    PolicySummaryExtended:
      type: object
      properties:
        id:
          type: string
          format: uuid
        policyNumber:
          type: string
        policyType:
          type: string
        status:
          type: string
        effectiveDate:
          type: string
          format: date
        expiryDate:
          type: string
          format: date
        premium:
          type: number
          format: float
        vehicle:
          $ref: '#/components/schemas/VehicleSummary'
    VehicleSummary:
      type: object
      properties:
        registrationNumber:
          type: string
        make:
          type: string
        model:
          type: string
        year:
          type: integer
    PurchasePolicyRequest:
      type: object
      required:
        - quoteId
        - paymentInfo
        - beneficiary
      properties:
        quoteId:
          type: string
          format: uuid
        customerId:
          type: string
          format: uuid
        paymentInfo:
          $ref: '#/components/schemas/PaymentInfo'
        beneficiary:
          $ref: '#/components/schemas/BeneficiaryInfo'
        additionalOptions:
          $ref: '#/components/schemas/AdditionalOptions'
    PaymentInfo:
      type: object
      required:
        - paymentOption
        - paymentMethod
        - paymentDetails
      properties:
        paymentOption:
          type: string
          enum: [FULL_PAYMENT, INSTALLMENT_6, INSTALLMENT_12]
        paymentMethod:
          type: string
          enum: [CREDIT_CARD, BANK_TRANSFER, PROMPTPAY, E_WALLET]
        paymentDetails:
          type: object
          properties:
            cardToken:
              type: string
              example: "encrypted-card-token"
            cardLast4:
              type: string
              example: "1234"
            cardType:
              type: string
              example: "VISA"
    BeneficiaryInfo:
      type: object
      properties:
        type:
          type: string
          enum: [SELF, SPOUSE, PARENT, CHILD, OTHER]
        relationship:
          type: string
        personalInfo:
          $ref: '#/components/schemas/PersonalInfo'
    AdditionalOptions:
      type: object
      properties:
        autoRenewal:
          type: boolean
          default: false
        digitalDocuments:
          type: boolean
          default: true
        smsNotifications:
          type: boolean
          default: true
    PurchasePolicyResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          properties:
            policyId:
              type: string
              format: uuid
            policyNumber:
              type: string
              example: "POL-2024-567890"
            status:
              type: string
              example: "ACTIVE"
            effectiveDate:
              type: string
              format: date-time
            expiryDate:
              type: string
              format: date-time
            premium:
              type: object
              properties:
                totalPremium:
                  type: number
                  format: float
                paymentStatus:
                  type: string
                paymentMethod:
                  type: string
                transactionId:
                  type: string
            documents:
              type: object
              properties:
                policyDocument:
                  type: string
                  format: uri
                certificate:
                  type: string
                  format: uri
                taxInvoice:
                  type: string
                  format: uri
            nextPaymentDue:
              type: string
              format: date
            message:
              type: string
              example: "Policy purchased successfully. Documents sent via email."
        timestamp:
          type: string
          format: date-time
...
    RenewPolicyRequest:
      type: object
      properties:
        renewalOptions:
          type: object
          properties:
            coverageChanges:
              $ref: '#/components/schemas/CoverageOptions'
            paymentPlan:
              type: string
              enum: [FULL_PAYMENT, INSTALLMENT_6, INSTALLMENT_12]
        paymentMethod:
          $ref: '#/components/schemas/PaymentInfo'
    RenewPolicyResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          properties:
            renewedPolicyId:
              type: string
              format: uuid
            renewedPolicyNumber:
              type: string
            effectiveDate:
              type: string
              format: date-time
            expiryDate:
              type: string
              format: date-time
            premium:
              type: object
              properties:
                newTotalPremium:
                  type: number
                  format: float
                noClaimDiscount:
                  type: number
                  format: float
                finalPremium:
                  type: number
                  format: float
            message:
              type: string
              example: "Policy renewed successfully with updated coverage"
        timestamp:
          type: string
          format: date-time
    CancelPolicyRequest:
      type: object
      required:
        - cancellationReason
        - effectiveDate
      properties:
        cancellationReason:
          type: string
          enum: [VEHICLE_SOLD, MOVED_ABROAD, FINANCIAL_REASONS, DISSATISFIED, OTHER]
        effectiveDate:
          type: string
          format: date
        documents:
          type: object
          properties:
            proofOfSale:
              type: string
              format: uuid
    CancelPolicyResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          properties:
            cancellationId:
              type: string
              format: uuid
            effectiveDate:
              type: string
              format: date-time
            refund:
              type: object
              properties:
                refundAmount:
                  type: number
                  format: float
                refundMethod:
                  type: string
                processingTime:
                  type: string
            message:
              type: string
              example: "Policy cancellation processed. Refund will be processed within 5-7 business days."
        timestamp:
          type: string
          format: date-time

    # Claims Management Schemas
...
    AdminPolicyListResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          properties:
            policies:
              type: array
              items:
                $ref: '#/components/schemas/AdminPolicySummary'
            pagination:
              $ref: '#/components/schemas/Pagination'
        timestamp:
          type: string
          format: date-time
    AdminPolicySummary:
      type: object
      properties:
        id:
          type: string
          format: uuid
        policyNumber:
          type: string
        customerName:
          type: string
        vehicleInfo:
          type: string
        policyType:
          type: string
        premium:
          type: number
          format: float
        status:
          type: string
        effectiveDate:
          type: string
          format: date
        expiryDate:
          type: string
          format: date
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
        size:
          type: integer
        totalElements:
          type: integer
        totalPages:
          type: integer
        hasNext:
          type: boolean
        hasPrevious:
          type: boolean
...